Refactor InspectableObject and CamerController
Simplified Inspectable object even more!! Now it seems to be working like a charm :heart: No need for stupid dual camera setup anymore.
Also added inspectionActive variable to CameraController so it is up to date if object is being inspected. No more errors on console!! :+1:

Removed InspectionCamera from scene, as it is not required anymore.